# Application Insights logger allocates less