--- a/src/Microsoft.Azure.WebJobs.Shared/DictionaryExtensions.cs
+++ b/src/Microsoft.Azure.WebJobs.Shared/DictionaryExtensions.cs
@@ -14,5 +14,15 @@ namespace System.Collections.Generic
             }
             return default(T);
         }
+
+        public static T GetValueOrDefault<T>(this IReadOnlyDictionary<string, object> dictionary, string key)
+        {
+            object value;
+            if (dictionary != null && dictionary.TryGetValue(key, out value))
+            {
+                return (T)value;
+            }
+            return default(T);
+        }
     }
 }