--- a/test/Microsoft.Azure.WebJobs.Host.UnitTests/PublicSurfaceTests.cs
+++ b/test/Microsoft.Azure.WebJobs.Host.UnitTests/PublicSurfaceTests.cs
@@ -232,6 +232,7 @@ namespace Microsoft.Azure.WebJobs.Host.UnitTests
                 "ParameterLog",
                 "Poco",
                 "RecoverableException",
+                "ReadOnlyScopeDictionary",
                 "ScopeKeys",
                 "SingletonOptions",
                 "TraceEvent",