… src/Microsoft.Azure.WebJobs.Host/Properties/AssemblyInfo.cs
… src/Microsoft.Azure.WebJobs.Logging.ApplicationInsights/Properties/AssemblyInfo.cs
… test/Benchmarks/WebJobsLoggingApplicationInsights.cs