This PR changes the logging with Application Insights. It does it in the following way.

1. `DictionaryLoggerScope` captures and persists the merged dictionary of values. With the changes in this PR it caches the value and reuses it, reducing allocations whenever one or more `.Log` operations are performed in a single `Scope`. A rough estimate is 2-3KB per call. 
1. The internal seam of `DictionaryLoggerScope` is changed to return `IReadOnlyDictionary<string,object>` to ensure that the caller does not interfere with the cached value.
1. A new `ReadOnlyScopeDictionary` is provided for using with `.BeginScope` where the keys are known and the only difference is in the array of values provided. For reasonably small number of keys brute-force compare should be more than enough.

The following changes were considered but not applied because of the lack of evidence of perf improvements

- prunning one LINQ call
- replacing an array `.Contains` with a `HashSet.Contains`

One thing to notice is that this PR makes the evaluation lazy and keeps the `IEnumerable<KeyValuePair<string,object>>` in the scope until it's used for the first time. On one side it has a positive aspect of lazy evaluation. If nothing is used, scope doesn't have to be evaluated. 

This PR does not change the underlying implementation to support `ISupportExternalScope` and `IExternalScopeProvider`. The external scope provided is controlled... externally 😅 and can have an impact on sematics of scoping. This should be considered as a separate PR.

### Benchmarks

Benchmarking two cases: a call to `.LogFunctionResult` and 3 subsequent calls to `.LogInformation`. The benchmarks were included in this PR as they test a broad area of API and might be a base for future improvements.

#### Before

|       Method |     Mean |    Error |   StdDev |   Median |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|---------:|-------:|-------:|------:|----------:|
|  OnException | 69.62 us | 0.813 us | 0.761 us | 69.53 us | 6.9580 | 0.1221 |     - |   21.3 KB |
| MultipleLogs | 44.36 us | 1.726 us | 4.979 us | 42.08 us | 5.1880 |      - |     - |  16.07 KB |

#### After

|       Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
|  OnException | 62.56 us | 1.209 us | 1.241 us | 4.8828 |      - |     - |  15.34 KB |
| MultipleLogs | 34.17 us | 0.656 us | 0.701 us | 3.2959 | 0.0610 |     - |  10.17 KB |

Again, the time improvements are low, but trimming a few KB per log statement looks quite good.