@brettsam 2020-10-09
Nice! This is definitely a hot area for perf improvements. This code is called a *lot* so any gains are multiplied across many invocations.

Is it possible to try a benchmark using the built-in `LoggerExternalScopeProvider`? This didn't exist when I first wrote this code, but I've moved to using it just about everywhere else (https://github.com/Azure/azure-functions-host/pull/5114).

Ideally I'd like to remove the `DictionaryLoggerScope` type completely but I took a dependency on it [here](https://github.com/Azure/azure-webjobs-sdk/blob/1dcece0045ff4de5fdddce6c587f047ad85eedef/src/Microsoft.Azure.WebJobs.Logging.ApplicationInsights/Initializers/WebJobsTelemetryInitializer.cs#L58) that we'd have to work through.

Side note -- I'm on-call right now so a fully review probably won't happen until next week :-)