@Scooletz 2020-10-09
This was my aim. Logging is called a lot and even when I can't reduce the price of App Insights, I could definitely make it cheaper on the execution side. 

Let me give it a spin in the next week then. I was not aware of the history behind it. I observed moving to `IExternalScopeProvider` in other repos but didn't consider it here. I think that it may a bit more complex, but will give it a try.