@Scooletz 2020-10-10
☝️  This will break a build. I'm in the middle of adding `IExternalScopeProvider` but due to external control of the scope I have some problems with making it "just build (TM)". Will try to fix it soon.