@Scooletz 2020-10-10
I'm not sure if I can provide a meaningful implementation based on `IExternalScopeProvider` as the scopes as described in a comment are needed in additional places. What should I do with it? Undo `IExternalScopeProvider` related work and leave it for a future rework? Some input will be highly appreciated.