@Scooletz 2020-10-19
Hey. Any input on https://github.com/Azure/azure-webjobs-sdk/pull/2603#issuecomment-706564583 @brettsam ?