@Scooletz 2020-10-20
FYI @mathewc @brettsam I removed the unfinished work with `IExternalScopeProvider` and `ISupportExternalScope` and focused on improving the existing one. As you can see, the amount of memory being saved in here is huge as I trimmed down allocs by 10KB per benchmark. Additionally, there should be some gain in `FunctionExecutor` as it uses the `ReadOnlyScopeDictionary` now as well.

I did my best to describe the reason for abandoning the `IExternalScopeProvider` work. This would require a huge discussion about owning scopes, using them as a merge-dictionary for Application Insights logger etc., something that would take much longer and would require multiple roundtrips IMO.

I'm looking forward to hearing from you, what do you think about the results.