@Scooletz 2020-10-20
I needed to rewrite the history. The reason for this is that I made a huge mistake in the initial benchmark as I took the call creating `FunctionInstance` as a part of the measurement. To fix it I amended the first commit with the benchmark, rerun it, rebased all my work on top of it and then rerun it again. Now it compares apples to apples.