@Scooletz 2020-10-20
> Yes, I agree that removing that ScopeProvider will take more work. I'm really regretting that I took that dependency the way I did.

I'm very opinionated in regards to the `IExternalScopeProvider` and who should be responsible for keeping scopes so I do not share the regret. At the same time, I do understand the overall approach for moving toward using it.

> Do the measurements in the initial PR comment now have the updated benchmark results?

Yes they do. I found that the allocs for the improved versions might vary (by 500bytes or so), but haven't found the reason.  They are left in the `Benchmarks` project so they may be rerun for the first commit as well as the last.

Sorry, for the red build. I'm running a subset of tests and this was the reason. The last one should be good.