@Scooletz 2020-10-24
I think I'll need to think a bit more about the approach. Eventually, all the scopes should use [LoggerMessage.DefineScope](https://docs.microsoft.com/en-us/dotnet/api/microsoft.extensions.logging.loggermessage) rather than passing some dictionaries. Maybe it's really about splitting the responsibilities of scopes? Some of them are used to navigate execution of the `ApplincationInsightsLogger` when other deliver properties for logging.

I think I'll need to pause for a while and think more about it.